Update to clang-tidy version 10

+++ tesseract/tesseract_collision/include/tesseract_collision/test_suite/collision_octomap_mesh_unit.hpp
@@ -106,14 +106,11 @@ inline void runTest(DiscreteContactManager& checker, const std::string& file_pat
       idx = 1;
 
     mesh_vertices_color[static_cast<std::size_t>(
-        (*mesh_triangles)[static_cast<std::size_t>(4 * r.subshape_id[static_cast<std::size_t>(idx)] + 1)])] =
-        Eigen::Vector3i(255, 0, 0);
+        (*mesh_triangles)[4 * r.subshape_id[static_cast<std::size_t>(idx)] + 1])] = Eigen::Vector3i(255, 0, 0);
     mesh_vertices_color[static_cast<std::size_t>(
-        (*mesh_triangles)[static_cast<std::size_t>(4 * r.subshape_id[static_cast<std::size_t>(idx)] + 2)])] =
-        Eigen::Vector3i(255, 0, 0);
+        (*mesh_triangles)[4 * r.subshape_id[static_cast<std::size_t>(idx)] + 2])] = Eigen::Vector3i(255, 0, 0);
     mesh_vertices_color[static_cast<std::size_t>(
-        (*mesh_triangles)[static_cast<std::size_t>(4 * r.subshape_id[static_cast<std::size_t>(idx)] + 3)])] =
-        Eigen::Vector3i(255, 0, 0);
+        (*mesh_triangles)[4 * r.subshape_id[static_cast<std::size_t>(idx)] + 3])] = Eigen::Vector3i(255, 0, 0);
   }
 
   writeSimplePlyFile(file_path, *mesh_vertices, mesh_vertices_color, *mesh_triangles, mesh->getTriangleCount());
